Only rerender progress bar instead of WHOLE modal
